smb: move some duplicate definitions to common/smbacl.h
[ Upstream commit b51174da743b6b7cd87c02e882ebe60dcb99f8bf ]

In order to maintain the code more easily, move duplicate definitions
to new common header file.

Acked-by: Namjae Jeon <linkinjeon@kernel.org>
Stable-dep-of: d07b26f39246 ("ksmbd: require minimum ACE size in smb_check_perm_dacl()")
